pkg/factory: split to mulitple files (noop)
Just break factory.go in to multiple files, noop otherwise.


Conflicts:
	go-controller/pkg/factory/factory.go
